+++ scripts/generate_patches.py
@@ -1,14 +1,7 @@
 from argparse import ArgumentParser, Namespace
-from pathlib import Path
 
-import pandas as pd
-from pandas import DataFrame
-
-import source.satellite_dataset.config as sdcfg
-import source.satellite_dataset.table as sd_table
+import source.patches.generation as p_generation
 import source.satellite_dataset.utility as sd_util
-import user.config as ucfg
-from source.satellite_dataset.typing import ImgGeoDataArrays
 
 
 def main() -> None:
@@ -26,29 +19,9 @@ def main() -> None:
     if patch_resolution is None:
         patch_resolution = int(input("Enter pixel resolution of patches: "))
 
-    dataset_archive = dataset["archive"]
-    dataset_path = Path(dataset["path"])
-
-    table_path = sdcfg.DATASET_TABLES_DIR_PATH / f"{dataset_name}.pkl"
-
-    if regenerate_table or not table_path.is_file():
-        sd_table.generate_dataset_table(dataset_archive, dataset_path, table_path)
-
-    dataset_table = pd.read_pickle(table_path)
-
-    # Spatial resolution of a patch in m/px (ignoring projection effects / distortions)
-    patch_resolution_mpx = patch_scale_km * 1000 / patch_resolution
-
-    match dataset_archive:
-        case "vex-vmc" | "vco":
-            generate_img_geo_patches(
-                dataset_archive, dataset_table, patch_resolution_mpx
-            )
-        case _:
-            raise ValueError(
-                "No patch generation script implemented for dataset archive "
-                f"'{dataset_archive}'"
-            )
+    p_generation.generate_patches(
+        dataset, dataset_name, patch_scale_km, patch_resolution, regenerate_table
+    )
 
 
 def parse_input_args() -> Namespace:
@@ -71,67 +44,6 @@ def parse_input_args() -> Namespace:
     )
 
     return arg_parser.parse_args()
-
-
-def generate_img_geo_patches(
-    dataset_archive: str, dataset_table: DataFrame, patch_resolution_mpx: float
-) -> None:
-    for row_index, row_data in dataset_table.iterrows():
-        img_max_resolution_mpx: float = row_data["max_resolution_mpx"]
-
-        if not passes_resolution_threshold(
-            img_max_resolution_mpx, patch_resolution_mpx
-        ):
-            continue
-
-        img_file_path = Path(row_data["img_file_path"])
-        geo_file_path = Path(row_data["geo_file_path"])
-
-        img_geo_data_arrays = load_img_geo_data_arrays(
-            dataset_archive, img_file_path, geo_file_path
-        )
-
-
-def load_img_geo_data_arrays(
-    dataset_archive: str, img_file_path: Path, geo_file_path: Path
-) -> ImgGeoDataArrays:
-    match dataset_archive:
-        case "vex-vmc":
-            raise NotImplementedError()
-        case "vco":
-            img_hdu = sd_util.load_fits_hdu_or_hdus(img_file_path, 1)
-            lat_hdu, lon_hdu, lt_hdu, ina_hdu, ema_hdu = sd_util.load_fits_hdu_or_hdus(
-                geo_file_path,
-                [
-                    "Latitude",
-                    "Longitude",
-                    "Local time",
-                    "Incidence angle",
-                    "Emission angle",
-                ],
-            )
-
-            data_arrays: ImgGeoDataArrays = {
-                "image": img_hdu.data,
-                "latitude": lat_hdu.data,
-                "longitude": lon_hdu.data,
-                "local_time": lt_hdu.data,
-                "incidence_angle": ina_hdu.data,
-                "emission_angle": ema_hdu,
-            }  # type: ignore
-
-            return data_arrays
-        case _:
-            raise ValueError(
-                "Can not load data arrays for unknown dataset archive "
-                f"{dataset_archive}"
-            )
-
-
-def passes_resolution_threshold(
-    img_max_resolution: float, patch_resolution: float
-) -> bool:
-    return img_max_resolution / patch_resolution < ucfg.PATCH_RESOLUTION_TOLERANCE
 
 
 if __name__ == "__main__":
